#463 delete file from s3 when deleted from site

+++ muckrock/foia/signals.py
@@ -6,7 +6,8 @@ from boto.s3.connection import S3Connection
 
 from muckrock.foia.models import FOIARequest, FOIAFile
 from muckrock.foia.tasks import upload_document_cloud
-from muckrock.settings import AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, AWS_STORAGE_BUCKET_NAME
+from muckrock.settings import AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, AWS_STORAGE_BUCKET_NAME,\
+                              AWS_DEBUG, DEBUG
 
 def foia_update_embargo(sender, **kwargs):
     """When embargo has possibly been switched, update the document cloud permissions"""
@@ -33,12 +34,14 @@ def foia_file_delete_s3(sender, **kwargs):
     """Delete file from S3 after the model is deleted"""
     # pylint: disable=unused-argument
 
-    foia_file = kwargs['instance']
+    if not DEBUG or AWS_DEBUG:
+        # only delete if we are using s3
+        foia_file = kwargs['instance']
 
-    conn = S3Connection(AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY)
-    bucket = conn.get_bucket(AWS_STORAGE_BUCKET_NAME)
-    key = bucket.get_key(foia_file.ffile.name)
-    key.delete()
+        conn = S3Connection(AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY)
+        bucket = conn.get_bucket(AWS_STORAGE_BUCKET_NAME)
+        key = bucket.get_key(foia_file.ffile.name)
+        key.delete()
 
 
 pre_save.connect(foia_update_embargo, sender=FOIARequest,
